+++ services/app-web/src/components/SubmissionSummarySection/SubmissionTypeSummarySection/SubmissionTypeSummarySection.tsx
@@ -4,7 +4,7 @@ import { ProgramT } from '../../../common-code/domain-models'
 import { DataDetail } from '../../../components/DataDetail'
 import { DoubleColumnGrid } from '../../../components/DoubleColumnGrid'
 import { SectionHeader } from '../../../components/SectionHeader'
-import { SubmissionTypeRecord } from '../../../constants/submissions'
+import { SubmissionTypeRecord } from '../../../constants/healthPlanPackages'
 import { DraftSubmission, StateSubmission } from '../../../gen/gqlClient'
 import { isStateSubmission } from '../../../gqlHelpers'
 import { usePreviousSubmission } from '../../../hooks/usePreviousSubmission'
